feat: impede de deletar tema ativo

+++ app/Controllers/Http/Temas/Mui/DeleteTemasController.ts
@@ -1,5 +1,6 @@
 import type { HttpContextContract } from "@ioc:Adonis/Core/HttpContext";
 import Database from "@ioc:Adonis/Lucid/Database";
+import ApiError from "App/Exceptions/ApiError";
 import TemaMuiSistema from "App/Models/TemaMuiSistema";
 import UploadImagem from "App/Services/UploadImagem";
 
@@ -9,6 +10,13 @@ export default class DeleteTemasController {
     await Database.transaction(async (trx) => {
       const tema = await TemaMuiSistema.findOrFail(id);
       tema.useTransaction(trx);
+
+      if (tema.ativo) {
+        throw new ApiError(
+          "O tema está atualmente em uso. Ative outro tema antes de deletar o atual.",
+          400
+        );
+      }
 
       await tema.delete();
